Improved enemy collision on attack
When the enemy is hit or grappled to, the enemy's hitbox will add the player to exclude. This is an improvement of before so grappling isn't immediately stopped by collision and, in some cases with the prior method, the enemy could get pushed upward (while gravity is disabled) and play the dying animation in air.